baseline: snapshot ECHO/agentic work before refactor
Captures the pre-refactor state so every later step is revertible:
- 18 core files modified (+497/-21)
- new core files: utils/env_loss.py, utils/env_pack.py,
  backends/sglang_utils/qwen35_hf_config.py, runtime_hooks/
- slime_plugins/agentic/ (61 files, 6333 lines)

main stays at upstream 681b3ad so 'git diff main' is the exact
local delta against upstream.

+++ slime/backends/megatron_utils/model_provider.py
@@ -83,7 +83,12 @@ class LinearForLastLayer(torch.nn.Linear):
         runtime_gather_output: bool | None = None,
     ) -> tuple[torch.Tensor, None]:
         logits = super().forward(input_)
-        logits = logits.float()
+        # Keep the outer [T, V] logits in bf16 end-to-end (2026-08-27, OOM fix):
+        # the fp32 copy doubled forward logits AND backward logits-grad memory
+        # (32.5 GiB each at 32k tokens x 248320 vocab) and OOMed H200-140G on
+        # near-full-context samples. The PG loss path (_VocabParallelLogProbEntropy)
+        # upcasts each 2048-token chunk to fp32 internally, so log-prob numerics
+        # are unchanged (upcast of the same bf16 GEMM output is bit-exact).
         if self.sequence_parallel:
             logits = tensor_parallel.gather_from_sequence_parallel_region(logits, tensor_parallel_output_grad=False)
         return logits, None
